2041: separate the ACAA functionality from the ACA (aca.go -> acaa.go)
